requirements.txt
selenium
pytest
webdriver-manager
python.dotenv
